pyproject.toml
[project]
name = "llm-azure"
version = "2.0"
description = "Azure AI Foundry plugin for LLM - access Anthropic, Llama, Mistral and more via Azure"
readme = "README.md"
authors = [{name = "Fabian Geiger"}]
license = "Apache-2.0"
requires-python = ">=3.10"
classifiers = []
dependencies = [
    "llm>=0.22",
    "azure-identity",
]

[project.optional-dependencies]
anthropic = ["anthropic"]

[project.urls]
Homepage = "https://github.com/fabge/llm-azure"
Changelog = "https://github.com/fabge/llm-azure/releases"
Issues = "https://github.com/fabge/llm-azure/issues"
CI = "https://github.com/fabge/llm-azure/actions"

[project.entry-points.llm]
azure = "llm_azure"

[dependency-groups]
dev = ["pytest"]

[tool.uv]
package = true

[tool.setuptools]
packages = ["llm_azure"]
